Make the benchmarks run for a predictable time.
Before this patch, some benchmarks would take much longer to run than
others, and the iteration count argument would often need to be
adjusted to find a reasonable compromise.

Now, the command-line argument specifies the desired run-time, so that
the same benchmarks can be used on all platforms.

Argument parsing and result printing are now centralised, to reduce
duplication, and to ensure that we use a consistent interface for future
benchmarks.

Change-Id: Ic4dd481402b2e556dadffef66f414af7ff3bd504