Fix the 'sh' field for ADD/SUB immediate.
We treated this as a two-bit field <23:22>, with encodings recognised
only when bit 23 was zero. The latest XML describes it as a one-bit
field <22>. The decoder already excludes encodings with bit 23 set, so
no behaviour is changed, except that the instruction masks (e.g.
`AddSubImmediateFMask`) should work more consistently.

This patch also renames `ShiftAddSub` to `ImmAddSubShift` to make it
clear that it applies to "immediate" forms, and not to "shifted" form.

Change-Id: Ieea2fc5c8a7cfb057983b7063372ee6f760d82a1